Refactoring the creation of the Space
The spectrum were wrongly sorted and then associated with the wrong filenames. This is now fixed. However there is a bug that results in the wrong order by which the Filenames are presented in the report window. It should be WR - VEG - WR2 - DC. It appears to be random now in the order.