Implemented chunked deletion
When deleting hierarchies we previously tried to delete all relevant rows in one go. For one hierarchy there are about 100k eav_ids and about 10 to 25k spectrum_ids this lead to massive delays in the inno-db engine. Based on http://mysql.rjweb.org/doc.php/deletebig this patch implements a chunked deletion of the relevant rows in packets or chunks of 1000 rows.

Improves inno-db engine performance and thus also the deletion process. Still takes a couply of minutes for a day of data.